Make the decorator and ExpirationDatabaseDriver internal to ExpirableKeySqlMap
