--- src/Take.Elephant.Sql/DbConnectionDecorator.cs
@@ -1,71 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Common;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Take.Elephant.Sql
-{
-    public sealed class DbConnectionDecorator : DbConnection
-    {
-        private readonly DbConnection _dbConnection;
-        public DbConnectionDecorator(DbConnection dbConnection)
-        {
-            _dbConnection = dbConnection;
-        }
-
-        public override string ConnectionString { get => _dbConnection.ConnectionString; set => _dbConnection.ConnectionString = value; }
-
-        public override string Database => _dbConnection.Database;
-
-        public override string DataSource => _dbConnection.DataSource;
-
-        public override string ServerVersion => _dbConnection.ServerVersion;
-
-        public override ConnectionState State => _dbConnection.State;
-
-        public override void ChangeDatabase(string databaseName)
-        {
-            _dbConnection.ChangeDatabase(databaseName);
-        }
-
-        public override void Close()
-        {
-            _dbConnection.Close();
-        }
-
-        public override void Open()
-        {
-            _dbConnection.Open();
-        }
-
-        protected override DbCommand CreateDbCommand()
-        {
-            var command = _dbConnection.CreateCommand();
-            var parameter = command.CreateParameter();
-            parameter.ParameterName = "@ExpirableKeySqlMap_ExpirationDate";
-            parameter.Value = DateTimeOffset.UtcNow;
-
-            command.Parameters.Add(parameter);
-            return command;
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
-            _dbConnection.Dispose();
-        }
-
-        public override async ValueTask DisposeAsync()
-        {
-            await base.DisposeAsync();
-            await _dbConnection.DisposeAsync();
-        }
-
-        protected override DbTransaction BeginDbTransaction(IsolationLevel isolationLevel)
-        {
-            return _dbConnection.BeginTransaction(isolationLevel);
-        }
-    }
-}
