+++ src/Take.Elephant.Sql/ExpirableKeySqlMap.cs
@@ -3,10 +3,12 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 using Take.Elephant.Sql.Mapping;
 
+[assembly: InternalsVisibleTo("Take.Elephant.Tests")]
 namespace Take.Elephant.Sql
 {
     /// <summary>
@@ -107,7 +109,7 @@ namespace Take.Elephant.Sql
         /// Injects a SQL statement to avoid retrieving expired items.
         /// </summary>
         /// <seealso cref="Take.Elephant.Sql.IDatabaseDriver" />
-        private class ExpirationDatabaseDriver : IDatabaseDriver
+        internal class ExpirationDatabaseDriver : IDatabaseDriver
         {
             private readonly IDatabaseDriver _underlyingDatabaseDriver;
             private readonly string _expirationColumnName;
@@ -119,7 +121,7 @@ namespace Take.Elephant.Sql
             }
 
             public DbConnection CreateConnection(string connectionString)
-                => new DbConnectionDecorator(_underlyingDatabaseDriver.CreateConnection(connectionString));
+                => new ExpirationDbConnection(_underlyingDatabaseDriver.CreateConnection(connectionString));
 
             public string GetSqlStatementTemplate(SqlStatement sqlStatement)
             {
@@ -159,6 +161,71 @@ namespace Take.Elephant.Sql
                 if (sql.Contains("ORDER BY")) return sql.Replace("ORDER BY", $"{filter} ORDER BY");
                 return $"{sql} {filter}";
             }
+        }
+    }
+
+    /// <summary>
+    /// Implements a new behavior for DbConnection with a custom parameter.
+    /// </summary>
+    internal class ExpirationDbConnection : DbConnection
+    {
+        private readonly DbConnection _dbConnection;
+        public ExpirationDbConnection(DbConnection dbConnection)
+        {
+            _dbConnection = dbConnection;
+        }
+
+        public override string ConnectionString { get => _dbConnection.ConnectionString; set => _dbConnection.ConnectionString = value; }
+
+        public override string Database => _dbConnection.Database;
+
+        public override string DataSource => _dbConnection.DataSource;
+
+        public override string ServerVersion => _dbConnection.ServerVersion;
+
+        public override ConnectionState State => _dbConnection.State;
+
+        public override void ChangeDatabase(string databaseName)
+        {
+            _dbConnection.ChangeDatabase(databaseName);
+        }
+
+        public override void Close()
+        {
+            _dbConnection.Close();
+        }
+
+        public override void Open()
+        {
+            _dbConnection.Open();
+        }
+
+        protected override DbCommand CreateDbCommand()
+        {
+            var command = _dbConnection.CreateCommand();
+            var parameter = command.CreateParameter();
+            parameter.ParameterName = "@ExpirableKeySqlMap_ExpirationDate";
+            parameter.Value = DateTimeOffset.UtcNow;
+
+            command.Parameters.Add(parameter);
+            return command;
+        }
+
+        protected override void Dispose(bool disposing)
+        {
+            base.Dispose(disposing);
+            _dbConnection.Dispose();
+        }
+
+        public override async ValueTask DisposeAsync()
+        {
+            await base.DisposeAsync();
+            await _dbConnection.DisposeAsync();
+        }
+
+        protected override DbTransaction BeginDbTransaction(IsolationLevel isolationLevel)
+        {
+            return _dbConnection.BeginTransaction(isolationLevel);
         }
     }
 }
